backend/requirements.txt
# Web Framework
fastapi==0.109.2
starlette==0.36.3
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
aiosqlite==0.19.0
psycopg==3.1.18

# Data Validation
pydantic-settings==2.2.1
pydantic[email]==2.6.1
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.1

# Rate Limiting
limits==4.0.1
slowapi==0.1.9

# Testing
pytest==8.0.2
httpx==0.27.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Build Tools
wheel==0.42.0
setuptools==69.1.1

# Redis
redis==5.2.1

# Postgres
psycopg2-binary==2.9.10

# FAISS
faiss-cpu==1.10.0

# Elasticsearch
elasticsearch==8.17.1

# Machine Learning
scikit-learn==1.6.1
pandas==2.2.3
sentence-transformers==3.4.1
torch==2.6.0
transformers==4.49.0
torchvision==0.21.0
torchaudio==2.6.0

pymupdf==1.25.3
python-docx==1.1.2
pytesseract==0.3.13
python-pptx==1.0.2

langdetect==1.0.9

asyncpg==0.30.0
greenlet==3.1.1